requirements.txt
python-pptx==0.6.21
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
